Add archive command for archiving data files
It doesn't do much; it just saves the current state to a user-specified file.

By default, it saves to a file named "(iso local date time).json", where
"(iso local date time)" is an ISO 8601 combined date and time representation
with second-level precision and no timezone specifier.

If a relative path is specified, it saves to a file with the given name, to the
same directory as the configured addressbook data path.

Finally, if an absolute path is specified, it saves to the given path.