Make Job and Money classes more robust and flexible
Instead of specifying a fixed per-hour rate, we specify a `Rate`, which
allows us to better describe pay rates (e.g. $20 per 4 hours, etc).

We use `Duration`s instead of `float`s for durations in `Job`s to avoid
floating point imprecision.

We also avoid the use of float operations in `Money` to force users of
the `Money` class to specify precise values (i.e. with BigDecimal).

Finally, we also add a (hopefully unique) job ID to `Job` instances, so
we can uniquely identify jobs with only the ID (since it's plausible
that there might be multiple jobs with the exact same attributes).